requirements.txt
torch==2.1.1
tqdm==4.22.0
seaborn==0.9.0
pandas==0.23.0
dill==0.2.8.2
torchvision==0.16.0
numpy==1.26.2
scipy==1.3.1
matplotlib==3.8.2
Pillow==10.1.0
scikit_learn==1.3.2
PyYAML==5.3
robustness
